Do not remove the file and disable the service during upgrade

+++ config/RPM/centos_postuninstall.sh
@@ -10,8 +10,11 @@ then
     rm -f /etc/init.d/${SERVICE}
 elif grep -q -i "release 7" /etc/redhat-release
 then
-    systemctl disable ${SERVICE}
-    rm -f /etc/systemd/system/${SERVICE}.service
+    if [ "$1" = "0" ]; then
+        # Prepare task for an uninstallation
+        systemctl disable ${SERVICE}
+        rm -f /etc/systemd/system/${SERVICE}.service
+    fi
     systemctl daemon-reload
 fi
